test: extract OpensearchIT, ElasticsearchIT from IncidentUpdateRepositoryIT
Extracted from tIncidentUpdateRepositoryIT because static inner classes
are not run from maven (but are run from IDEs)
